+++ modules/sandbox/src/sandbox.h
@@ -675,8 +675,8 @@ private:
 	void initialize_syscalls();
 	static void initialize_syscalls_2d();
 	static void initialize_syscalls_3d();
-	GuestVariant *setup_arguments(gaddr_t &sp, const Variant **args, int argc);
-	void setup_arguments_native(gaddr_t arrayDataPtr, GuestVariant *v, const Variant **args, int argc);
+	GuestVariant *setup_arguments(gaddr_t &sp, const Variant **args, int64_t argc);
+	void setup_arguments_native(gaddr_t arrayDataPtr, GuestVariant *v, const Variant **args, int64_t argc);
 
 	machine_t *m_machine = nullptr;
 	Node *m_tree_base = nullptr;
